blabla bla

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -23,7 +23,7 @@ int get_sizeof(unsigned long int n)
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int k = 1;
+	unsigned long int k = 1;
 	int len = get_sizeof(n);
 
 	if (len > 0)
